fix: Disable batch execute button if there is a pending tx, move tx functions to util file
